[tests] preserve existing attention test logic
AI assistance was used for this change.
